Create LargerActuator.f3z
Shared the Large  actuator Fusion 360 file so others can take a look. Currently the elbow cap is complete, the main shell is missing the PCB holder, the RGB lght holder, and the screw holes to connect the stepper plate. This is a workin progress and not yet ready for printing.